test: move script related tests to separate module
