+++ cli/tests/it/main.rs
@@ -8,11 +8,18 @@ mod config;
 #[cfg(not(feature = "external-integration-tests"))]
 mod create;
 #[cfg(not(feature = "external-integration-tests"))]
+mod script;
+#[cfg(not(feature = "external-integration-tests"))]
 mod test_cmd;
 #[cfg(not(feature = "external-integration-tests"))]
 mod utils;
 #[cfg(not(feature = "external-integration-tests"))]
 mod verify;
+
+// import forge utils as mod
+#[allow(unused)]
+#[path = "../../src/utils.rs"]
+pub(crate) mod forge_utils;
 
 #[cfg(feature = "external-integration-tests")]
 mod integration;
