+++ cli/tests/it/verify.rs
@@ -1,7 +1,6 @@
 //! Contains various tests for checking forge commands related to verifying contracts on etherscan
 
 use crate::utils::{self, EnvExternalities};
-
 use foundry_cli_test_utils::{
     forgetest,
     util::{TestCommand, TestProject},
@@ -135,7 +134,7 @@ fn verify_watch_on_chain(info: Option<EnvExternalities>, prj: TestProject, mut c
             "--watch".to_string(),
             address,
             contract_path.to_string(),
-            info.etherscan.to_string(),
+            info.etherscan,
         ]);
         parse_verification_result(&mut cmd, 6).expect("Failed to verify check")
     }
@@ -281,7 +280,7 @@ forgetest!(
                     "--rpc-url".to_string(),
                     info.rpc.clone(),
                     "--private-key".to_string(),
-                    info.pk.clone(),
+                    info.pk,
                 ])
                 .arg("--broadcast")
                 .arg("--verify")
